Optimizes task and primogems bot display and command

+++ src/main/javascript/Command/Task/taskCommand.js
@@ -7,7 +7,7 @@ commands.add = async function(msg, prefix, keyword) {
     const newMsg = msg.content.replace(prefix + keyword, "").trim();
 
     if (newMsg === '') {
-        msg.reply("Please indicate a task to be saved.");
+        msg.reply("Please indicate a task to be saved with `!taskadd <task name>`.");
     } else {
         await taskSchema.create({
             message: newMsg,
@@ -37,7 +37,7 @@ commands.add = async function(msg, prefix, keyword) {
             .addField("Successfully saved task:", newMsg)
             .addField("Here are list of tasks you have saved:", desc)
             .setColor(0xF1C40F)
-        msg.reply(embed);
+        msg.channel.send(embed);
     }
 }
 
@@ -68,7 +68,7 @@ commands.list = async function(msg) {
             .setTitle("Task Command: List")
             .addField("Here are list of tasks you have saved:", desc)
             .setColor(0xF1C40F)
-        msg.reply(embed);
+        msg.channel.send(embed);
     }
 }
 
@@ -96,14 +96,15 @@ commands.delete = async function(msg, prefix, keyword) {
 
                 desc += completed + " " + (i+1) + ": " + list[i].message + "\n";
             }
-    
+            
+            msg.reply("Please reply with the number to delete, for example `2` to delete the second item in your task list. To cancel command, type `cancel`.")
             let filter = m => m.author.id === msg.author.id
             const embed = new Discord.RichEmbed()
                 .setAuthor(msg.author.username + "#" + msg.author.discriminator + "'s Task List", msg.author.avatarURL)
                 .setTitle("Task Command: Delete")
-                .addField("Here is your list of task. Please reply with the number to delete:", desc)
+                .addField("Here is your list of task:", desc)
                 .setColor(0xF1C40F)
-            msg.reply(embed).then(() => {
+            msg.channel.send(embed).then(() => {
             msg.channel.awaitMessages(filter, {
                 max: 1,
                 time: 10000,
@@ -115,7 +116,7 @@ commands.delete = async function(msg, prefix, keyword) {
                     if (!isNaN(message)) {
                         const num = parseInt(message.content)-1;
 
-                        if (parseInt(message.content) > 0 && parseInt(message.content) <= list.length) {
+                        if (num+1 > 0 && num+1 <= list.length) {
                             const taskToDelete = list[num];
                             await taskSchema.deleteOne(list[num]);
                             msg.reply("Successfully deleted task: `" + taskToDelete.message + "`.");
@@ -123,7 +124,11 @@ commands.delete = async function(msg, prefix, keyword) {
                             msg.reply("Please redo the command `" + prefix + keyword + "` again, then insert a proper number within the task list.")
                         }
                     } else {
-                        msg.reply("Please redo the command `" + prefix + keyword + "` again, then insert a number.")
+                        if (message.content.split(" ")[0].toLowerCase() === 'cancel') {
+                            msg.reply("`" + prefix + keyword + "` command cancelled.");
+                        } else {
+                            msg.reply("Please redo the command `" + prefix + keyword + "` again, then insert a number.");
+                        }
                     }
                     
                 })
@@ -135,7 +140,7 @@ commands.delete = async function(msg, prefix, keyword) {
     }
 }
 
-commands.clear = async function(msg) {
+commands.clear = async function(msg, prefix, keyword) {
     const list = await taskSchema.find({
         author: msg.author.id
     })
@@ -143,8 +148,35 @@ commands.clear = async function(msg) {
     if (list.length === 0) {
         msg.reply("You do not have any saved task.");
     } else {
-        await taskSchema.deleteMany({});
-        msg.reply("Successfully deleted all tasks saved.");
+        let filter = m => m.author.id === msg.author.id
+        msg.reply("Are you sure you want to delete all your saved tasks? Initiating this action is irreversible. Reply `yes` to delete all, reply `no` to cancel.").then(() => {
+            msg.channel.awaitMessages(filter, {
+                max: 1,
+                time: 10000,
+                errors: ['time']
+                })
+                .then(async message => {
+                    message = message.first();
+                    const receivedMsg = message.content.split(" ")[0];
+
+                    if (receivedMsg.toLowerCase() === 'yes') {
+                        await taskSchema.deleteMany({});
+                        msg.reply("Successfully deleted all tasks saved.");
+                    } else if (receivedMsg.toLowerCase() === 'no') {
+                        msg.reply("`" + prefix + keyword + "` command cancelled.");
+                    } else {
+                        msg.reply("Please redo the command `" + prefix + keyword + "` again if you wish to delete all your saved tasks.");
+                    }
+
+                }).catch(collected => {
+                    console.log(collected);
+                    msg.reply("Timeout. Please redo the command again.");
+                });
+        })
+
+
+
+        
     }
     
 }
@@ -176,15 +208,16 @@ commands.edit = async function(msg, prefix, keyword) {
     
             let filter = m => m.author.id === msg.author.id
 
+            msg.reply("Please reply with the number to edit followed by content to be edited, for example `2 do homework`. To cancel command, type `cancel`.")
             const embed = new Discord.RichEmbed()
                 .setAuthor(msg.author.username + "#" + msg.author.discriminator + "'s Task List", msg.author.avatarURL)
                 .setTitle("Task Command: Edit")
-                .addField("Here is your list of task. Please reply with the number to edit followed by content to be edited, for example `2 do homework`:", desc)
+                .addField("Here is your list of task:", desc)
                 .setColor(0xF1C40F)
-            msg.reply(embed).then(() => {
+            msg.channel.send(embed).then(() => {
             msg.channel.awaitMessages(filter, {
                 max: 1,
-                time: 10000,
+                time: 60000,
                 errors: ['time']
                 })
                 .then(async message => {
@@ -192,7 +225,7 @@ commands.edit = async function(msg, prefix, keyword) {
                     const num = parseInt(message.content.split(" ")[0])-1;
 
                     if (!isNaN(num)) {
-                        if (parseInt(message.content) > 0 && parseInt(message.content) <= list.length) {
+                        if (num+1 > 0 && num+1 <= list.length) {
                             const newMsg = message.content.replace(message.content.split(" ")[0] + " ", "").trim();
 
                             if (message.content.split(" ").length < 2) {
@@ -215,7 +248,11 @@ commands.edit = async function(msg, prefix, keyword) {
                             msg.reply("Please redo the command `" + prefix + keyword + "` again, then insert a proper number within the task list and ensure that it is `<task number> <message to edit>`, for example `2 do homework`.")
                         }
                     } else {
-                        msg.reply("Please redo the command `" + prefix + keyword + "` again, then ensure it is `<task number> <message to edit>`, for example `2 do homework`.")
+                        if (message.content.split(" ")[0].toLowerCase() === 'cancel') {
+                            msg.reply("`" + prefix + keyword + "` command cancelled.")
+                        } else {
+                            msg.reply("Please redo the command `" + prefix + keyword + "` again, then ensure it is `<task number> <message to edit>`, for example `2 do homework`.")
+                        }
                     }
                 })
                 .catch(collected => {
@@ -250,14 +287,16 @@ commands.markComplete = async function(msg, prefix, keyword) {
 
                 desc += completed + " " + (i+1) + ": " + list[i].message + "\n";
             }
-    
+            
+
+            msg.reply("Please reply with the number to mark task as complete. To cancel command, type `cancel`.")
             let filter = m => m.author.id === msg.author.id
             const embed = new Discord.RichEmbed()
                 .setAuthor(msg.author.username + "#" + msg.author.discriminator + "'s Task List", msg.author.avatarURL)
                 .setTitle("Task Command: Mark Complete")
-                .addField("Here is your list of task. Please reply with the number to mark task as complete:", desc)
+                .addField("Here is your list of task:", desc)
                 .setColor(0xF1C40F)
-            msg.reply(embed).then(() => {
+            msg.channel.send(embed).then(() => {
             msg.channel.awaitMessages(filter, {
                 max: 1,
                 time: 10000,
@@ -268,8 +307,7 @@ commands.markComplete = async function(msg, prefix, keyword) {
                     
                     if (!isNaN(message)) {
                         const num = parseInt(message.content)-1;
-
-                        if (parseInt(message.content) > 0 && parseInt(message.content) <= list.length) {
+                        if (num+1 > 0 && num+1 <= list.length) {
                             const task = list[num].message;
 
                             await taskSchema.updateOne(list[num], {
@@ -307,7 +345,11 @@ commands.markComplete = async function(msg, prefix, keyword) {
                             msg.reply("Please redo the command `" + prefix + keyword + "` again, then insert a proper number within the task list.")
                         }
                     } else {
-                        msg.reply("Please redo the command `" + prefix + keyword + "` again, then insert a number.")
+                        if (message.content.split(" ")[0].toLowerCase() === 'cancel') {
+                            msg.reply("`" + prefix + keyword + "` command cancelled.");
+                        } else {
+                            msg.reply("Please redo the command `" + prefix + keyword + "` again, then insert a number.");
+                        }
                     }
                     
                 })
@@ -343,14 +385,15 @@ commands.markIncomplete = async function(msg, prefix, keyword) {
 
                 desc += completed + " " + (i+1) + ": " + list[i].message + "\n";
             }
-    
-            let filter = m => m.author.id === msg.author.id
+
+            msg.reply("Please reply with the number to mark task as incomplete. To cancel command, type `cancel`.");
+            let filter = m => m.author.id === msg.author.id;
             const embed = new Discord.RichEmbed()
                 .setAuthor(msg.author.username + "#" + msg.author.discriminator + "'s Task List", msg.author.avatarURL)
                 .setTitle("Task Command: Mark Complete")
-                .addField("Here is your list of task. Please reply with the number to mark task as complete:", desc)
+                .addField("Here is your list of task:", desc)
                 .setColor(0xF1C40F)
-            msg.reply(embed).then(() => {
+            msg.channel.send(embed).then(() => {
             msg.channel.awaitMessages(filter, {
                 max: 1,
                 time: 10000,
@@ -362,7 +405,7 @@ commands.markIncomplete = async function(msg, prefix, keyword) {
                     if (!isNaN(message)) {
                         const num = parseInt(message.content)-1;
 
-                        if (parseInt(message.content) > 0 && parseInt(message.content) <= list.length) {
+                        if (num+1 > 0 && num+1 <= list.length) {
                             const task = list[num].message;
 
                             await taskSchema.updateOne(list[num], {
@@ -400,7 +443,11 @@ commands.markIncomplete = async function(msg, prefix, keyword) {
                             msg.reply("Please redo the command `" + prefix + keyword + "` again, then insert a proper number within the task list.")
                         }
                     } else {
-                        msg.reply("Please redo the command `" + prefix + keyword + "` again, then insert a number.")
+                        if (message.content.split(" ")[0].toLowerCase() === 'cancel') {
+                            msg.reply("`" + prefix + keyword + "` command cancelled.");
+                        } else {
+                            msg.reply("Please redo the command `" + prefix + keyword + "` again, then insert a number.");
+                        }
                     }
                     
                 })
